refactored segmented visitor log code
