+++ plugins/Goals/templates/getGoalReportView.twig
@@ -38,7 +38,7 @@
         {% endif %}
     </ul>
 
-    <a href="javascript:;" class="segmentedlog" onclick="DataTable_RowActions_Registry.getActionByName('SegmentVisitorLog').createInstance({}).showVisitorLog('Goals.getMetrics', 'visitConvertedGoalId=={{ idGoal }}', {})">
+    <a href="javascript:;" class="segmentedlog" onclick="SegmentedVisitorLog.show('Goals.getMetrics', 'visitConvertedGoalId=={{ idGoal }}', {})">
         {{ 'Live_RowActionTooltipWithDimension'|translate('General_Goal'|translate) }}
     </a>
 {% endif %}
